Simplify util again now it only has a single usecase
Previous commits saw this util split so the sizing scheme could
base it's calculations on the number of gutters that were present. In
the previous commit we removed that as a cause of overrendering so we
can simplify the util again.
