--- packages/harpguru-core/src/hooks/use-octave-column-groups-fragment-aware/index.ts
@@ -1,1 +1,0 @@
-export { useOctaveColumnGroupsFragmentAware } from './use-octave-column-groups-fragment-aware'
