--- packages/harpguru-core/src/hooks/use-octave-column-groups-fragment-aware/use-octave-column-groups-fragment-aware.ts
@@ -1,11 +1,0 @@
-import { useGlobal } from 'reactn'
-
-import { useOctaveColumnGroups } from '../use-octave-column-groups'
-import type { ColumnRanges } from '../use-octave-column-groups'
-
-export const useOctaveColumnGroupsFragmentAware = (): ColumnRanges => {
-  const [fragmentHarpFaceByOctaves] = useGlobal('fragmentHarpFaceByOctaves')
-  const octaveColumnGroups = useOctaveColumnGroups()
-  if (fragmentHarpFaceByOctaves) return octaveColumnGroups
-  return [octaveColumnGroups.flat()]
-}
