+++ packages/harpguru-core/src/hooks/use-octave-column-groups/use-octave-column-groups.test.ts
@@ -8,9 +8,25 @@ jest.mock('reactn')
 const mockUseGlobal = useGlobal as jest.Mock
 
 test('Identifies the columns grouped by octave', () => {
-  mockUseGlobal.mockImplementation(buildMockUseGlobalImplementation({}))
+  mockUseGlobal.mockImplementation(
+    buildMockUseGlobalImplementation({
+      fragmentHarpFaceByOctaves: true,
+    })
+  )
   const octaveColumnGroups = useOctaveColumnGroups()
   const expectedOctaveGroups = [[0], [1, 2, 3, 4], [5, 6, 7], [8, 9]]
+
+  expect(octaveColumnGroups).toStrictEqual(expectedOctaveGroups)
+})
+
+test('Returns an ungrouped set if the harp fragmentation is off', () => {
+  mockUseGlobal.mockImplementation(
+    buildMockUseGlobalImplementation({
+      fragmentHarpFaceByOctaves: false,
+    })
+  )
+  const octaveColumnGroups = useOctaveColumnGroups()
+  const expectedOctaveGroups = [[0, 1, 2, 3, 4, 5, 6, 7, 8, 9]]
 
   expect(octaveColumnGroups).toStrictEqual(expectedOctaveGroups)
 })
@@ -19,6 +35,7 @@ test('Returns a bounded grouped set', () => {
   mockUseGlobal.mockImplementation(
     buildMockUseGlobalImplementation({
       sourceColumnBounds: [2, 7],
+      fragmentHarpFaceByOctaves: true,
     })
   )
   const octaveColumnGroups = useOctaveColumnGroups()
@@ -26,6 +43,19 @@ test('Returns a bounded grouped set', () => {
     [2, 3, 4],
     [5, 6, 7],
   ]
+
+  expect(octaveColumnGroups).toStrictEqual(expectedOctaveGroups)
+})
+
+test('Returns a bounded ungrouped set', () => {
+  mockUseGlobal.mockImplementation(
+    buildMockUseGlobalImplementation({
+      sourceColumnBounds: [3, 8],
+      fragmentHarpFaceByOctaves: false,
+    })
+  )
+  const octaveColumnGroups = useOctaveColumnGroups()
+  const expectedOctaveGroups = [[3, 4, 5, 6, 7, 8]]
 
   expect(octaveColumnGroups).toStrictEqual(expectedOctaveGroups)
 })
