+++ packages/harpguru-core/src/hooks/use-octave-column-groups/use-octave-column-groups.ts
@@ -9,24 +9,24 @@ import type { ColumnRanges } from './get-octave-column-groups'
 import { arrayHasRoot } from './array-has-root'
 
 export const useOctaveColumnGroups = (): ColumnRanges => {
-  const [degreeMatrix] = useGlobal('activeDegreeMatrix')
+  const [viewableDegreeMatrices] = useGlobal('viewableDegreeMatrix')
   const [columnBounds] = useGlobal('columnBounds')
+  const [fragmentHarpFaceByOctaves] = useGlobal('fragmentHarpFaceByOctaves')
 
-  const columnsFirstDegreeMatrix = transposeMatrix(
-    extractHarpFaceFacts(degreeMatrix, 'harpface1')
-  ) as HarpFaceMatrix<Degree>
-  const rootColumnsMask = columnsFirstDegreeMatrix.map(arrayHasRoot)
-  const octaveColumnGroups = getOctaveColumnGroups(rootColumnsMask)
+  const octaveColumnGroups = (() => {
+    const columnsFirstDegreeMatrix = transposeMatrix(
+      extractHarpFaceFacts(viewableDegreeMatrices, 'harpface1')
+    ) as HarpFaceMatrix<Degree>
+    const rootColumnsMask = columnsFirstDegreeMatrix.map(arrayHasRoot)
+    const unshiftedColumnGroups = getOctaveColumnGroups(rootColumnsMask).filter(
+      (group) => group.length > 0
+    )
+    if (columnBounds == 'FIT') return unshiftedColumnGroups
+    return unshiftedColumnGroups.map((group) =>
+      group.map((index) => index + columnBounds[0])
+    )
+  })()
 
-  if (columnBounds === 'FIT') return octaveColumnGroups
-
-  const [startColumn, endColumn] = columnBounds
-
-  const viewableColumnGroups = octaveColumnGroups
-    .map((group) => {
-      return group.filter((index) => index >= startColumn && index <= endColumn)
-    })
-    .filter((group) => group.length > 0)
-
-  return viewableColumnGroups
+  if (fragmentHarpFaceByOctaves) return octaveColumnGroups
+  return [octaveColumnGroups.flat()]
 }
